+++ src/server/mod.rs
@@ -6,9 +6,11 @@ mod datachan;
 mod failed_logins;
 pub(crate) mod ftpserver;
 mod password;
+#[cfg(feature = "proxy_protocol")]
 mod proxy_protocol;
 mod session;
 pub(crate) mod shutdown;
+mod switchboard;
 mod tls;
 
 pub(crate) use chancomms::ControlChanMsg;